pyproject.toml
[project]
name = "mcp-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ffmpeg>=1.4",
    "loguru>=0.7.3",
    "mistralai>=2.2.0",
    "pyradio>=0.5.2",
    "python-dotenv>=1.2.2",
    "python-ffmpeg>=2.0.12",
    "pyyaml>=6.0.3",
    "strands-agents-tools>=0.3.0",
    "strands-agents[anthropic]>=1.34.1",
    "strands-deepgram>=0.1.0",
]
